+++ src/main/java/com/example/db_project/repositories/LuggageRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LuggageRepository extends JpaRepository<Luggage, Integer> {
+public interface LuggageRepository extends JpaRepository<Luggage, Long> {
 }
